+++ docs/changelog.rst
@@ -26,6 +26,7 @@ Fixed
 ^^^^^
 - Fixed DataAdapter.resolve_paths with unknown keys #121
 - Fixed the WGS84 datum in the gis_utils.utm_crs method.
+- In merge.merge the grid is now aligned with input dataset with the largest overlap if no dst_bounds & dst_res are given.
 
 Deprecated
 ^^^^^^^^^^